+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Note that since we don't clearly distinguish between a public and private interf
 - Add `defaultSnapshotIndex` argument to `MVSLoadOptions` to enable loading a snapshot other than the first one by default
 - Fix `loaders.loadMvs*` options type
 - Read simulation box from gro and lammps files
+- Rework Superposition UI panel
+- Add Ligand alignments by maximum common connected subgraphs (MCCS)
 - Handle unobserved residues from `entity_poly_seq` (#965)
 
 ## [v5.10.0] - 2026-06-14